fix patch versioning
When committing a bug fix directly to qa branch, do not
increment the patch version right away.  This way, one
can do git log to get the commit hash for the fix and
apply it as a cherry-pick on the develop branch without
changing any version numbers in develop.  After the cherry-pick,
go back to qa AND THEN change and commit the patch version.